requirements.txt
langchain
openai
streamlit
pandas
seaborn
matplotlib
faiss-cpu
python-dotenv